Web: Bundling bugfixed
It turned out Bundling did not work with IIS Express at all. In Debug mode, it is disabled; but in Release mode, it should have been enabled. Now, it wasn't.
Now testing with IIS revealed this; because in IIS with Release mode, the bundling was really activated - and it was configured buggy, as it turned out
Bugs:
* There was a bundle called "Content/themes/main". It was an incorrect name, because this is also a valid path of a directory in the web app. So there was a collision --> GET requests were interpreted for the bundle as directory browsing requests
** So renamed the bundle (to "Content/themes/MainStyles")
* There were urls with relative path eg in this file: "Content/themes/main/Site.css". Since after bundling its "path" became "Content/themes/MainStyles" (so 1 dir upper), the relative urls got broken